Rename 'model' to a repository.
