+++ tests/units/Models/CampaignTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Models\Campaign;
+use App\Repositories\CampaignRepository;
 
 class CampaignTest extends TestCase
 {
@@ -9,7 +9,8 @@ class CampaignTest extends TestCase
     {
         // Setup a mock collection of campaigns!
 
-        $campaigns = Campaign::getAll();
+        $campaign = $this->app->make(CampaignRepository::class);
+        $campaigns = $campaign->getAll();
 
         $this->assertNotCount(0, $campaigns);
     }
@@ -18,8 +19,7 @@ class CampaignTest extends TestCase
     public function can_get_a_campaign_by_slug()
     {
         // Set up a mock campaign!
-
-        $campaign = Campaign::findBySlug('baby-its-cold-inside');
+        $campaign = app(CampaignRepository::class)->findBySlug('baby-its-cold-inside');
 
         $this->assertEquals('Baby, It\'s Cold Inside', $campaign->getTitle());
     }
